backend/requirements.txt
# ============================================
# 核心框架
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# ============================================
# 认证与安全
# ============================================
# JWT 认证
python-jose[cryptography]==3.3.0
# 密码加密
passlib[bcrypt]==1.7.4
# OAuth2 支持
python-multipart==0.0.6

# ============================================
# 数据处理与验证
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# 视频处理
# ============================================
# OpenCV (视频流处理、图像增强)
opencv-python==4.8.1.78
# NumPy (数组操作、LUT 映射)
numpy==1.24.3

# ============================================
# 网络与 HTTP
# ============================================
# HTTP 客户端
httpx==0.25.1
# CORS 支持（FastAPI 内置，但列出以明确）
# python-multipart 已包含

# ============================================
# 开发工具（可选）
# ============================================
# 代码格式化
# black==23.11.0
# 代码检查
# flake8==6.1.0
# 类型检查
# mypy==1.7.0

# ============================================
# 测试工具（可选）
# ============================================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# httpx (已在上方列出，用于测试 API)